Fold MegaphoneSelection into Trackblazer as a nested object
